Have added cards project, also changed variables a bit
Variables used to all be floats but this is not applicable in all cases, so have made most objects other than stats. Have also removed IHasAssetCode as in most cases this can be expressed as a variable if needed and can be made more specific for each use case this way.