Adding tag functionality to OpenRpg
This will not be apparently useful straight away, but when more procedural related libraries come in it will be useful to be able to use tags as a way to describe procedural content.

It can also be used for other content too, but for the moment its just something in preparation for later.