Altering Inventory and adding basic IEntity
There has always been the notion of an entity but it has been at a higher level but this has caused issues when things are pulling in too many directions, i.e a monster doesnt need race/class in all scenarios but a character would, so this represents a common base level, the 2 (or more) can inherit from.

Also updated Inventory to remove HasItem method as it can be represented agnostically of implementation and have more overloads as an extension method.